% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPSMKL-package.R
\docType{package}
\name{BPSMKL-package}
\alias{BPSMKL}
\alias{BPSMKL-package}
\title{BPSMKL package}
\description{
Bayesian pathway sparse multiple kernel learning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yuzi896/BPSMKL}
  \item Report bugs at \url{https://github.com/Yuzi896/BPSMKL/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
